another update

+++ WeatherApp/Models/Root.cs
@@ -39,7 +39,7 @@ namespace WeatherApp.Models
 		public Clouds clouds { get; set; }
 		public Wind wind { get; set; }
 		public int visibility { get; set; }
-		public int pop { get; set; }
+		public double pop { get; set; }
 		public Sys sys { get; set; }
 		public string dt_txt { get; set; }
 	}
